mybaits

+++ src/main/java/life/majiang/community/controller/PublishController.java
@@ -49,7 +49,7 @@ public class PublishController {
     public String doPblish(@RequestParam("title") String title,
                            @RequestParam(value = "description", required = false) String description,
                            @RequestParam(value = "tag", required = false) String tag,
-                           @RequestParam(value = "id", required = false) Long id,
+                           @RequestParam(value = "id", required = false) Integer id,
                            HttpServletRequest request,
                            Model model){
